src/ShitpostBot.MlService/requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
sentence-transformers==5.1.2
paddleocr==3.3.2
paddlepaddle==3.2.2
opencv-python-headless==4.12.0.88
pillow==10.1.0
python-magic==0.4.27
python-multipart==0.0.20
transformers==4.57.3
pytesseract==0.3.13
